refactor(startup): move multi-channel preflight command handlers
Refs #933
