+++ crates/tau-startup/src/lib.rs
@@ -20,8 +20,10 @@ use tau_gateway::{
 use tau_session::validate_session_file;
 
 pub mod startup_model_catalog;
+pub mod startup_multi_channel_commands;
 
 pub use startup_model_catalog::*;
+pub use startup_multi_channel_commands::*;
 
 pub trait StartupPreflightActions {
     fn execute_onboarding_command(&self, cli: &Cli) -> Result<()>;
